requirements.txt
opencv-python==4.11.0.86
opencv-contrib-python==4.11.0.86
mediapipe==0.10.21
pandas==2.2.3
numpy==1.26.4
pillow==11.2.1
matplotlib==3.10.1